cheerio is tricky
When setting `xml` option to true, cheerio assumes that
it parses an XML document. This affects serialization:
all empty tags (like `<script src=""></script>`) become
self-closing (`<script src=""/>`), and this breaks browsers.

To overcome this, we need to manipulate on the raw HTML,
not on the one generated by `$.html()`.

To do that, we ask cheerio to include `startIndex` and `endIndex`
into the Node object. Then, when we need to save the result,
we check if there is an existing CSP tag. If yes, we cut it off
using `startIndex` and `endIndex`. After that we just add
the resulting `<meta>` tag right after `<head>`.